Consolidated veto logic.
