+++ packages/stateful/hooks/index.ts
@@ -14,6 +14,7 @@ export * from './useLoadingDaoCardInfos'
 export * from './useLoadingGovProposal'
 export * from './useMembership'
 export * from './useProposalPolytoneState'
+export * from './useProposalVetoState'
 export * from './useQuerySyncedRecoilState'
 export * from './useSimulateCosmosMsgs'
 export * from './useSyncWalletSigner'
