+++ packages/stateful/proposal-module-adapter/adapters/DaoProposalMultiple/components/ProposalStatusAndInfo.tsx
@@ -9,19 +9,14 @@ import {
   RotateRightOutlined,
   Send,
   Tag,
-  ThumbDownOutlined,
 } from '@mui/icons-material'
 import clsx from 'clsx'
-import { useRouter } from 'next/router'
 import { ComponentType, useCallback, useEffect, useState } from 'react'
 import toast from 'react-hot-toast'
 import { useTranslation } from 'react-i18next'
-import { useRecoilValue, waitForAll } from 'recoil'
+import { useRecoilValue } from 'recoil'
 
-import {
-  DaoCoreV2Selectors,
-  DaoProposalMultipleSelectors,
-} from '@dao-dao/state'
+import { DaoProposalMultipleSelectors } from '@dao-dao/state'
 import {
   CopyToClipboardUnderline,
   IconButtonLink,
@@ -30,28 +25,19 @@ import {
   ProposalStatusAndInfoProps,
   ProposalStatusAndInfo as StatelessProposalStatusAndInfo,
   TooltipTruncatedText,
-  useCachedLoading,
   useConfiguredChainContext,
   useDaoInfoContext,
-  useDaoNavHelpers,
 } from '@dao-dao/stateless'
 import {
-  ActionKey,
   BaseProposalStatusAndInfoProps,
   CheckedDepositInfo,
   DepositRefundPolicy,
-  EntityType,
   ProposalStatusEnum,
 } from '@dao-dao/types'
 import { MultipleChoiceVote } from '@dao-dao/types/contracts/DaoProposalMultiple'
 import {
-  CHAIN_GAS_MULTIPLIER,
-  cwMsgToEncodeObject,
   formatPercentOf100,
-  getDaoProposalSinglePrefill,
   getProposalStatusKey,
-  makeCw1WhitelistExecuteMessage,
-  makeWasmMessage,
   processError,
 } from '@dao-dao/utils'
 
@@ -60,9 +46,9 @@ import { ButtonLink } from '../../../../components/ButtonLink'
 import {
   DaoProposalMultipleHooks,
   useAwaitNextBlock,
-  useEntity,
   useMembership,
   useProposalPolytoneState,
+  useProposalVetoState,
   useWallet,
 } from '../../../../hooks'
 import { useProposalModuleAdapterOptions } from '../../../react'
@@ -124,19 +110,13 @@ const InnerProposalStatusAndInfo = ({
   depositInfo: CheckedDepositInfo | undefined
 }) => {
   const { t } = useTranslation()
-  const router = useRouter()
   const {
     chain: { chain_id: chainId },
     config: { explorerUrlTemplates },
   } = useConfiguredChainContext()
   const { coreAddress } = useDaoInfoContext()
-  const { getDaoProposalPath } = useDaoNavHelpers()
   const { proposalModule, proposalNumber } = useProposalModuleAdapterOptions()
-  const {
-    isWalletConnected,
-    address: walletAddress = '',
-    getSigningCosmWasmClient,
-  } = useWallet()
+  const { isWalletConnected, address: walletAddress = '' } = useWallet()
   const { isMember = false } = useMembership({
     coreAddress,
   })
@@ -261,239 +241,13 @@ const InnerProposalStatusAndInfo = ({
     awaitNextBlock,
   ])
 
-  const vetoConfig = 'veto' in config ? config.veto : undefined
-  const vetoEnabled = !!vetoConfig
-  const [vetoLoading, setVetoLoading] = useState<
-    'veto' | 'earlyExecute' | false
-  >(false)
-  const vetoerEntity = useEntity(vetoConfig?.vetoer || '')
-  // Flatten vetoer entities in case a cw1-whitelist is the vetoer.
-  const vetoerEntities = !vetoerEntity.loading
-    ? vetoerEntity.data.type === EntityType.Cw1Whitelist
-      ? vetoerEntity.data.entities
-      : [vetoerEntity.data]
-    : []
-  const vetoerDaoEntities = vetoerEntities.filter(
-    (entity) => entity.type === EntityType.Dao
-  )
-  // This is the voting power the current wallet has in each of the DAO vetoers.
-  const walletDaoVetoerMemberships = useCachedLoading(
-    !vetoerEntity.loading && walletAddress
-      ? waitForAll(
-          vetoerDaoEntities.map((entity) =>
-            DaoCoreV2Selectors.votingPowerAtHeightSelector({
-              contractAddress: entity.address,
-              chainId,
-              params: [{ address: walletAddress }],
-            })
-          )
-        )
-      : undefined,
-    undefined
-  )
-  const canBeVetoed =
-    vetoEnabled &&
-    (statusKey === 'veto_timelock' ||
-      (statusKey === ProposalStatusEnum.Open && vetoConfig.veto_before_passed))
-  // Find matching vetoer for this wallet, which is either the wallet itself or
-  // a DAO this wallet is a member of. If a matching vetoer is found, this
-  // wallet can veto.
-  const matchingWalletVetoer =
-    canBeVetoed && !vetoerEntity.loading
-      ? vetoerEntities.find(
-          (entity) =>
-            entity.type === EntityType.Wallet &&
-            entity.address === walletAddress
-        ) ||
-        (!walletDaoVetoerMemberships.loading && walletDaoVetoerMemberships.data
-          ? vetoerDaoEntities.find(
-              (_, index) =>
-                walletDaoVetoerMemberships.data![index].power !== '0'
-            )
-          : undefined)
-      : undefined
-  const walletCanEarlyExecute =
-    !!matchingWalletVetoer &&
-    statusKey === 'veto_timelock' &&
-    !!vetoConfig?.early_execute
-  const onVeto = useCallback(async () => {
-    if (vetoerEntity.loading || !matchingWalletVetoer) {
-      return
-    }
-
-    setVetoLoading('veto')
-    try {
-      if (
-        vetoerEntity.data.type === EntityType.Wallet ||
-        (vetoerEntity.data.type === EntityType.Cw1Whitelist &&
-          matchingWalletVetoer.type === EntityType.Wallet)
-      ) {
-        const client = await getSigningCosmWasmClient()
-        const msg = makeWasmMessage({
-          wasm: {
-            execute: {
-              contract_addr: proposalModule.address,
-              funds: [],
-              msg: {
-                veto: {
-                  proposal_id: proposalNumber,
-                },
-              },
-            },
-          },
-        })
-
-        if (vetoerEntity.data.type === EntityType.Wallet) {
-          await client.signAndBroadcast(
-            walletAddress,
-            [cwMsgToEncodeObject(msg, walletAddress)],
-            CHAIN_GAS_MULTIPLIER
-          )
-        } else {
-          await client.signAndBroadcast(
-            walletAddress,
-            [
-              cwMsgToEncodeObject(
-                makeCw1WhitelistExecuteMessage(vetoerEntity.data.address, msg),
-                walletAddress
-              ),
-            ],
-            CHAIN_GAS_MULTIPLIER
-          )
-        }
-
-        await onVetoSuccess()
-      } else if (matchingWalletVetoer.type === EntityType.Dao) {
-        router.push(
-          getDaoProposalPath(matchingWalletVetoer.address, 'create', {
-            prefill: getDaoProposalSinglePrefill({
-              actions: [
-                {
-                  actionKey: ActionKey.VetoOrEarlyExecuteDaoProposal,
-                  data: {
-                    chainId,
-                    coreAddress,
-                    proposalModuleAddress: proposalModule.address,
-                    proposalId: proposalNumber,
-                    action: 'veto',
-                  },
-                },
-              ],
-            }),
-          })
-        )
-      }
-    } catch (err) {
-      console.error(err)
-      toast.error(processError(err))
-
-      // Stop loading if errored.
-      setVetoLoading(false)
-    }
-
-    // Loading will stop on success when status refreshes.
-  }, [
-    vetoerEntity,
-    matchingWalletVetoer,
-    onVetoSuccess,
-    proposalNumber,
-    getSigningCosmWasmClient,
-    walletAddress,
-    proposalModule.address,
-    router,
-    getDaoProposalPath,
-    chainId,
-    coreAddress,
-  ])
-  const onVetoEarlyExecute = useCallback(async () => {
-    if (vetoerEntity.loading || !matchingWalletVetoer) {
-      return
-    }
-
-    setVetoLoading('earlyExecute')
-    try {
-      if (
-        vetoerEntity.data.type === EntityType.Wallet ||
-        (vetoerEntity.data.type === EntityType.Cw1Whitelist &&
-          matchingWalletVetoer.type === EntityType.Wallet)
-      ) {
-        const client = await getSigningCosmWasmClient()
-        const msg = makeWasmMessage({
-          wasm: {
-            execute: {
-              contract_addr: proposalModule.address,
-              funds: [],
-              msg: {
-                execute: {
-                  proposal_id: proposalNumber,
-                },
-              },
-            },
-          },
-        })
-
-        if (vetoerEntity.data.type === EntityType.Wallet) {
-          await client.signAndBroadcast(
-            walletAddress,
-            [cwMsgToEncodeObject(msg, walletAddress)],
-            CHAIN_GAS_MULTIPLIER
-          )
-        } else {
-          await client.signAndBroadcast(
-            walletAddress,
-            [
-              cwMsgToEncodeObject(
-                makeCw1WhitelistExecuteMessage(vetoerEntity.data.address, msg),
-                walletAddress
-              ),
-            ],
-            CHAIN_GAS_MULTIPLIER
-          )
-        }
-
-        await onExecuteSuccess()
-      } else if (matchingWalletVetoer.type === EntityType.Dao) {
-        router.push(
-          getDaoProposalPath(matchingWalletVetoer.address, 'create', {
-            prefill: getDaoProposalSinglePrefill({
-              actions: [
-                {
-                  actionKey: ActionKey.VetoOrEarlyExecuteDaoProposal,
-                  data: {
-                    chainId,
-                    coreAddress,
-                    proposalModuleAddress: proposalModule.address,
-                    proposalId: proposalNumber,
-                    action: 'earlyExecute',
-                  },
-                },
-              ],
-            }),
-          })
-        )
-      }
-    } catch (err) {
-      console.error(err)
-      toast.error(processError(err))
-
-      // Stop loading if errored.
-      setVetoLoading(false)
-    }
-
-    // Loading will stop on success when status refreshes.
-  }, [
-    vetoerEntity,
-    matchingWalletVetoer,
-    getSigningCosmWasmClient,
-    proposalModule.address,
-    proposalNumber,
-    onExecuteSuccess,
-    walletAddress,
-    router,
-    getDaoProposalPath,
-    chainId,
-    coreAddress,
-  ])
+  const { vetoEnabled, canBeVetoed, vetoOrEarlyExecute, vetoInfoItems } =
+    useProposalVetoState({
+      statusKey,
+      vetoConfig: 'veto' in config ? config.veto : undefined,
+      onVetoSuccess,
+      onExecuteSuccess,
+    })
 
   const info: ProposalStatusAndInfoProps<MultipleChoiceVote>['info'] = [
     {
@@ -512,18 +266,7 @@ const InnerProposalStatusAndInfo = ({
         <EntityDisplay {...props} address={proposal.proposer} noCopy />
       ),
     },
-    // Show vetoers if can be vetoed or was vetoed. If was not vetoed but could
-    // have been, don't show because it might confuse the user and make them
-    // think it was vetoed.
-    ...(vetoConfig && (canBeVetoed || statusKey === ProposalStatusEnum.Vetoed)
-      ? (vetoerEntities.map((entity) => ({
-          Icon: ThumbDownOutlined,
-          label: t('title.vetoer'),
-          Value: (props) => (
-            <EntityDisplay {...props} address={entity.address} noCopy />
-          ),
-        })) as ProposalStatusAndInfoProps<MultipleChoiceVote>['info'])
-      : []),
+    ...vetoInfoItems,
     {
       Icon: RotateRightOutlined,
       label: t('title.status'),
@@ -712,18 +455,7 @@ const InnerProposalStatusAndInfo = ({
       }
       info={info}
       status={status}
-      vetoOrEarlyExecute={
-        matchingWalletVetoer
-          ? {
-              loading: vetoLoading,
-              onVeto,
-              onEarlyExecute: walletCanEarlyExecute
-                ? onVetoEarlyExecute
-                : undefined,
-              isVetoerDaoMember: matchingWalletVetoer.type === EntityType.Dao,
-            }
-          : undefined
-      }
+      vetoOrEarlyExecute={vetoOrEarlyExecute}
       vote={
         loadingWalletVoteInfo &&
         !loadingWalletVoteInfo.loading &&
